+++ Core/Inc/Measurement/AHT20/aht20_interface.h
@@ -6,7 +6,7 @@
 /* ----------------------------------------------------------------------------------- */
 
 void AHT20Init(void);
-int AHT20_GetTemp(void);
-int AHT20_GetHum(void);
+float AHT20_GetTemp(void);
+float AHT20_GetHum(void);
 
 #endif /* __AHT20_INTERFACE_H */
